requirements.txt
streamlit==1.37.1
pdfminer.six==20240706
python-docx==1.1.2
janome==0.5.0
rapidfuzz==3.9.6
beautifulsoup4==4.12.3
requests==2.32.3
langdetect==1.0.9
lxml==5.2.2
